feat: add shared tutorial creation hook
